# shared/types/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Inventory System API",
        "version": "0.0.1",
        "description": "Nigeria health supply-chain inventory API. Contract: docs/API_CONTRACT.md."
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/alerts/{alert}/acknowledge": {
            "post": {
                "operationId": "alert.acknowledge",
                "summary": "Mark an open alert as seen and being handled. No-op if already acknowledged",
                "tags": [
                    "Alert"
                ],
                "parameters": [
                    {
                        "name": "alert",
                        "in": "path",
                        "required": true,
                        "description": "The alert ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AlertResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AlertResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/alerts/{alert}/resolve": {
            "post": {
                "operationId": "alert.resolve",
                "summary": "Close an alert. If the stock situation still holds, it is not raised\nagain until it clears and recurs. No-op if already resolved",
                "tags": [
                    "Alert"
                ],
                "parameters": [
                    {
                        "name": "alert",
                        "in": "path",
                        "required": true,
                        "description": "The alert ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AlertResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AlertResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/facilities/{facility}/stock-activity": {
            "get": {
                "operationId": "stockActivity.facility",
                "summary": "Stock Card entries per week at one facility (received / issued / other)",
                "tags": [
                    "StockActivity"
                ],
                "parameters": [
                    {
                        "name": "facility",
                        "in": "path",
                        "required": true,
                        "description": "The facility ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "weeks": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "week_start": {
                                                                "anyOf": [
                                                                    {
                                                                        "type": "string"
                                                                    },
                                                                    {}
                                                                ]
                                                            }
                                                        },
                                                        "required": [
                                                            "week_start"
                                                        ],
                                                        "additionalProperties": {}
                                                    }
                                                },
                                                "totals": {
                                                    "type": "object",
                                                    "properties": {
                                                        "received": {
                                                            "type": "number"
                                                        },
                                                        "issued": {
                                                            "type": "number"
                                                        },
                                                        "other": {
                                                            "type": "number"
                                                        }
                                                    },
                                                    "required": [
                                                        "received",
                                                        "issued",
                                                        "other"
                                                    ]
                                                },
                                                "last_transaction_date": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "weeks",
                                                "totals",
                                                "last_transaction_date"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "AlertResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "alert_type": {
                        "$ref": "#/components/schemas/AlertType"
                    },
                    "severity": {
                        "$ref": "#/components/schemas/AlertSeverity"
                    },
                    "status": {
                        "$ref": "#/components/schemas/AlertStatus"
                    },
                    "condition_active": {
                        "type": "boolean",
                        "description": "False once the stock situation that raised the alert has cleared."
                    },
                    "facility": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "name"
                        ]
                    },
                    "product": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "sku": {
                                "type": "string"
                            },
                            "unit_of_measure": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "name",
                            "sku",
                            "unit_of_measure"
                        ]
                    },
                    "batch": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Set for expiry alerts only.",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "batch_no": {
                                "type": "string"
                            },
                            "expiry_date": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "batch_no",
                            "expiry_date"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "acknowledged_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "acknowledged_by_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "resolved_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "resolved_by_name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Null when resolved automatically because the condition cleared."
                    }
                },
                "required": [
                    "id",
                    "alert_type",
                    "severity",
                    "status",
                    "condition_active",
                    "facility",
                    "product",
                    "batch",
                    "created_at",
                    "acknowledged_at",
                    "acknowledged_by_name",
                    "resolved_at",
                    "resolved_by_name"
                ],
                "title": "AlertResource"
            },
            "AlertSeverity": {
                "type": "string",
                "enum": [
                    "info",
                    "warning",
                    "critical"
                ],
                "title": "AlertSeverity"
            },
            "AlertStatus": {
                "type": "string",
                "enum": [
                    "open",
                    "acknowledged",
                    "resolved"
                ],
                "title": "AlertStatus"
            },
            "AlertType": {
                "type": "string",
                "description": "Same values as alerts.alert_type in docs/DATABASE_SCHEMA.md.\n",
                "enum": [
                    "stock_out",
                    "expired",
                    "low_stock",
                    "expiring_soon"
                ],
                "title": "AlertType"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}